feat(database): title column icons support quick renaming
