+++ packages/blocks/src/database-block/components/column-header/styles.ts
@@ -68,16 +68,13 @@ export const styles = css`
   .affine-database-column-drag-handle.dragging::before {
     display: block;
   }
-  .affine-database-column-content:hover {
+  .affine-database-column-content:hover,
+  .affine-database-column-content.edit {
     background: linear-gradient(
         0deg,
         var(--affine-hover-color),
         var(--affine-hover-color)
       ),
-      var(--affine-white);
-  }
-  .affine-database-column-content.edit {
-    background: linear-gradient(0deg, rgba(0, 0, 0, 0.08), rgba(0, 0, 0, 0.08)),
       var(--affine-white);
   }
   .affine-database-column-text {
@@ -89,19 +86,54 @@ export const styles = css`
     font-size: 14px;
     font-weight: 600;
   }
-  .affine-database-column-text svg {
+  .affine-database-column-type-icon {
+    display: flex;
+    align-items: center;
+    border: 1px solid transparent;
+    border-radius: 4px;
+  }
+  .affine-database-column-type-icon.edit {
+    background: var(--affine-white);
+    border-color: var(--affine-border-color);
+  }
+  .affine-database-column-type-icon.active {
+    border-color: var(--affine-border-color);
+    background: linear-gradient(
+        0deg,
+        var(--affine-hover-color),
+        var(--affine-hover-color)
+      ),
+      var(--affine-white);
+  }
+  .affine-database-column-type-icon svg {
     width: 16px;
     height: 16px;
     fill: var(--affine-icon-color);
   }
-  .affine-database-column-text.select svg {
-    fill: none;
+  .affine-database-column-text-content {
+    flex: 1;
+    display: flex;
+    align-items: center;
+    margin-right: 6px;
+  }
+  .affine-database-column-content:hover .affine-database-column-text-icon {
+    opacity: 1;
   }
   .affine-database-column-text-input {
     flex: 1;
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
+  }
+  .affine-database-column-text-icon {
+    display: flex;
+    align-items: center;
+    width: 16px;
+    height: 16px;
+    background: var(--affine-white);
+    border: 1px solid var(--affine-border-color);
+    border-radius: 4px;
+    opacity: 0;
   }
   .affine-database-column-input {
     width: 100%;
